fix disagreement in the number of rounds
The equation for the number of rounds in `generate_mds.sage`
appeared to be an approximate form of the equations in section 5.5.2
describing the minimum number of rounds to resist algebraic attacks.
The comment is consistent with this, describing this function as a
minimum number of rounds as a sanity check.

I've replaced the approximation with the exact formula for the
S(x) = x^{\alpha} case, which results in a number of rounds that
agrees with what we got from `calc_round_numbers.py`.