add appendix G based tests
Appendix G in the paper contains several tables with concrete
instances of Poseidon for alpha={-1, 3, 5}. These are generated
using the Python script's `print_latex_table_combinations` function.

This commit adds tests based on the output from
`print_latex_table_combinations`.

All vectors agree between the two implementations, although I've modified
the Python code to use `log2(p)` instead of `n` to match the equations in the
paper (this does make a difference for small primes by 1 round or so, i.e.
within the security margin). The modified equations are the Grobner 1
constraint and the interpolation constraint.