feat: support matrices that are not symmetric
The MDS matrices generated using our poseidon-paramgen crate
should be symmetric. However, if we use matrices that are not
constructed to be symmetric (as was the case with the original
parameter set for poseidon377), we need to be careful to compute
transposes, instead of relying on M^T = M (symmetry).