feat(r1cs): use `CryptographicSpongeVar`, add optional r1cs feature
In this commit we refactor our ark-sponge tests, originally written
to check our poseidon-permutation implementation against the previous
ark-sponge implementation.

We need to use the `CryptographicSpongeVar` for R1CS (since the permute
methods are private), so we do this to ensure the out-of-
circuit use of the `CryptographicSpongeVar` API agrees with our
custom poseidon-permutation implementation.