test: delete out of date test from vendored ark-sponge
This test is failing in the forked repo, after the change in:

https://github.com/penumbra-zone/sponge/commit/113469aef48a9c5f458dbf523a509d4769affbf2

That change was previously made for consistency with our `poseidon377` test
vectors (after ark-sponge changed how they invoke the permutation in the
sponge). Our poseidon377 vectors are `crate::hash` in the tests
as well as in the protocol spec - those tests pass so there
should be no issues with consistency across implementations.