paramgen: Grobner basis round for v1, v2 (closes #45)
In [0] some Grobner basis attacks were presented that apply
either partially to the 384-bit and 512-bit security level and
fully to the 1024-bit security level. Instead of updating the
v1 round numbers for these higher security levels, we panic
to alert the user that the choice may not be secure.

For poseidon2, we implement the round selection including
the adjustments for the improved grobner basis attacks
in [0].

[0]: https://eprint.iacr.org/2023/537.pdf