paramgen: perron's irreducibility criterion (needed for m_i calc)
We need to be able to compute the minimal polynomial in order
to check that a randomly generated internal matrix is a secure choice,
i.e. that it does not have arbitrarily long subspace trails.

We need to be able to compute the minimal polynomial of a matrix
and check it's irreducible. We can't yet do the former, but with
Perron's criteron for monic polynomials we can do the latter with
this commit.